+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `ClientCache` into `ConnectedClients`.
+
 ## [0.23.0] - 2024-02-22
 
 ### Changed